note scripts: can have an empty line in a note if the next line starts with a space
An annoying problem when writing marking or diagnostic algorithms was
that you couldn't put an empty line in something like a big `let` call:

note_1:
    let(
        a, 1,
        b, 2,

        a+b
    )

note_2:
    note_1

This commit changes the regex so that the next thing after the two line
breaks must be a non-space character, i.e. the next line after the gap
must not be indented.

This allows gaps inside expressions like the one above, where the next
line after the gap is indented.