use MathJax 3
This is a big breaking change, so I think we need to rename the default
theme to something like `default_2024`.

The MathJax URL needs to change; maybe we need a new switch to pass a
MathJax v3 URL.

There are lots of themes building on the old default theme that assume
they'll use MathJax v2.