student_id and initial_seed are exam-level variables
The variables `student_id` and `initial_seed` are set in the exam scope
and available to all questions. They can be used to coordinate the
randomisation across questions.

The initial seed is either a value from the built-in random number
generator, or loaded from the SCORM element 'numbas.initial_seed'.
The LTI provider can use this to force attempts to use a particular
seed, so they match a printed copy, for example.

fixes #805