findvars on recursive custom functions doesn't include that function's name
When finding unbound variables inside the definition of a custom JME
function, the scope didn't yet know about the function being defined, so
a recursive call of that function would produce the function's own name
as a free variable.

This changes makeJMEFunction to add the function to the scope that is
used when finding unbound variables inside its definition.