tidy up time displays
All the displays of time now have text content and machine-readable
`datetime` attributes.

This adds a Numbas.display_util.duration_observable object which has
computed properties `display` and `machine`, giving string renderings of
a duration, represented as a number of seconds.

There are two new Knockout binding handlers, `datetime` and `duration`,
which set the text content and `datetime` attribute of a `<time>`
element.