Suspend data is compressed with gzip
see #1080

This changes ScormStorage.(get|set)SuspendData to use
`CompressionStream` and `DecompressionStream` to compress the suspend
data with gzip before saving it with the SCORM API.

Streams are async, so all the methods to do with loading suspend data
had to become async.

At the moment all of the exam unit tests pass, but it's possible I've
missed a race condition somewhere.

Since async/await is at over 99% compatibility, I think it would make
sense to go through and make lots more methods async, but that's a big
job and would be a breaking change.