CSS: use @layer and & nesting rules
This commit hugely reorganises the stylesheets for the default theme, to
use layers and the CSS nesting selector.

Hopefully, this will be more readable and it'll be easier to find the
right place for a rule.

The question CSS preamble is now in its own layer, so it takes
precedence over the base stylesheet without us having to prefix all its
rules.