In custom color scheme, keep the text contrast consistent.
When using a custom colour scheme, the text colour corresponding to
each of the colours chosen by the user aims to match the contrast
between the main background and text colours.

So if a user goes for a low contrast between background and text in the
main content, then the text on things like buttons and the sidebar will
have the same contrast level.