make custom functions in a scope aware of all other custom functions
In a setup with functions:

    f1() = f2(3)

    f2(x) = 1+random(1..x)

Then a variable `q = f1()` seemed to depend on a variable `f2` because
the custom findvars method for `f1` didn't know about `f2` when it was
made.

This commit changes `Numbas.jme.makeFunctions` so that first of
all a temporary scope is made where all the custom functions are
defined, and then the real definitions of those functions are made in
the scope that will be used.