Another go at fixing findvars on custom JME functions
There was another case that wasn't caught by the last fix to findvars on
custom JME functions: when the definition of `a` calls `b(c())`, and
`a`, `b` and `c` are all custom JME functions.

Because the list of external variables in a function's definition was
only computed once, if `a` is defined before `b` and `c` then it won't
know about `c`. Running the process twice only got us one step further
down the chain of dependencies.

This commit changes the findvarsOps function for a custom JME function
to find the list of external variables each time the function is called.
That could be a bit slow, so we could maybe set a flag on the scope each time a
new function is defined, but it's no different to running findvars on a
long expression anyway.