use a multiplication symbol when RHS is a negative number
The regex for the case when there's a real number on the LHS wasn't
checking if the RHS was a negative number, just if it started with a
digit.

Checking the string rendering of the RHS is a bit indirect, but it
articulates that you don't want a string of digits or + or - signs next
to each other.