Don't automatically show warnings as soon as they're made
fixes #1135

Previously, the warning message pop up was shown as soon as any warning
was created.

The only way for this to happen when you don't have an input focused is
as a result of submitting the part.

It's very easy for the warning box to overlap the part feedback or
other parts of the prompt. In gap-fill parts, there can be so many boxes
that almost the entire part is covered.

This changes the logic so that warnings are only shown when you focus
an input element.

"Invalid input" warnings are usually repeated as "fail" feedback
messages, so the student will still see why their answer was invalid
after submitting.