fix warning box positioning (again)
I noticed that the code to position warning boxes was using the
document's width to work out if warning boxes should be pushed left, but
now that the <main> element has overflow-x: hidden, it should be using the
width of the question element.

This changes the positioning code to do that, and I think it resolves
some errors in positioning warnings for gaps, so they shouldn't ever
overlap the input box now.