remove Sarissa; fix XSLT transforms in Firefox 137
Firefox 137 seems to have fixed a bug in XSLT which all browsers have
had, and which our code was relying on.

It seems that when applying an XSLT stylesheet to a document with
contextNode `<question>`, for example, then a template with selector
`question` won't match that context node, only ones underneath it.

This commit fixes this by cloning the XML to be transformed and
attaching it to a new container element, then transforming the
container.

I've also taken the opportunity to remove Sarissa and
jquery.xslTransform, since we don't need IE compatibility any more.

I've tested on Firefox 137, Chrome 133 and Safari 18.3.