don't put the exam XML in scripts.js
This changes the compiler so it doesn't include the exam XML in
scripts.js, but in a separate file called exam.xml.

The function `Numbas.init` loads this file. It can be given an object
defining the URL to load the exam XML from, or just with the exam XML as
a string.

Working towards #1140