The compiler doesn't make XML any more
I've rewritten the code that converts the .exam JSON to XML in
JavaScript, so the conversion is done at runtime.

The compiler is now just in charge of putting together all the necessary
files.

The list of extensions in the .exam file contained absolute paths for
the compiler to use, which shouldn't end up in the finished package. So
I've changed the editor to put just the short names in the exam data,
and pass the absolute paths as a command-line argument.

When the page loads, the source.exam file is fetched, the custom part
type and extension definitions are loaded from it, and then the exam
JSON is converted to XML for use by the theme templates.

The runtime Exam object is still loaded from the XML. I think we could
switch to using the JSON, and only use the XML for producing the
displayed HTML.

still working towards #1140