Only init exam storage after questions are ready
Now that questions can take some time to become ready after the list is
initialised, the total marks available for the exam aren't calculated at
the time Exam.init runs.

This moves the storage initialisation to happen after the questions are
all ready. I don't think that anything can use the storage in the
intervening time, so it should be safe.