If math.Fraction is given a non-integer, numerator and denominator are scaled up to integers
Someone reported that matrixmath.inverse was returning something like
the identity matrix for any 3×3 rotation matrix.

`matrixmath.inverse` works by converting the matrix to a matrix of
Fractions, then performing Gauss-Jordan elimination on them. The
Fraction constructor rounds off its inputs, because it's expecting
integers on top and bottom.

This commit changes the Fraction constructor so both numerator and
denominator are multiplied by powers of 2 until they're both integers.

This necessarily results in some inaccuracy due to floating point error,
but at least it's not completely wrong.

I tried explicitly writing out the routine for inverting 3×3 matrices,
but it turned out to be less accurate than this method. The floating
point error must accumulate because there are more float operations.