allowPrinting: false means nothing at all is printed
fixes #1100

There are three options that control printing: allowPrinting,
resultsprintquestions and resultsprintadvice.

If the last two options were turned off, allowPrinting didn't really
change anything.

So this commit changes `allowPrinting: false` to mean that no print
button should be displayed, and in the print stylesheet the whole
document is hidden.