fiddle with the question theme footer
I think we want the footer to take up as little space as possible, to
improve things when embedding single questions in a page.

This puts the "display options" button on the same line as the copyright
text, and adds the logo, since it's not otherwise shown in the question
theme.