starting to rearrange so multiple exams can run in the page
There is now a <numbas-exam> element, which contains all of the exam
display. The exam-specific display init code has been moved inside that,
so hopefully multiple exams can be embedded in the page.

I'd like to end up with the <numbas-exam> element being in charge of
loading the exam too, using a source URL from an attribute, or a script
tag inside it with the exam definition. Then we can package up a generic
runtime, and page authors can embed Numbas exams wherever just by
writing the element, and without any explicit JavaScript calls.

At the moment, it looks like the stylesheet loads OK, but MathJax
doesn't run inside the exam, and the question theme thinks it needs to
leave space for the sidebar.

Could the stylesheet only be loaded inside the numbas-exam element?