BIG refactor!
Lots of stuff has moved around.

The modals are now inside the <numbas-exam> element, and it also has its
own loading progress display. To make this work, a lot of calls to show
modals, the lightbox, and the progress display had to be moved within
the <numbas-exam> element, and a couple of remaining helpers in
Numbas.controls now take an `exam` argument.

The ExamDisplay object has a property `root_element` referring to the
`<numbas-exam>` element.

The exam now has its own task scheduler, a simple chain of promises
managed by a `Numbas.Scheduler` object. Really, we could get rid of
`Numbas.schedule.add`.