Finished fixing the worksheet theme?
The display options modal is now in the worksheet theme. I've moved the page margin inputs there.

The generator interface is forced to use the default styles, so the changes only affect the generated content.

I've moved each of the modals to their own template file, to make them easier to override.

The page margins are now set by a <style> tag with some dynamically generated CSS, so that they're applied to the @page selector, which doesn't inherit from the document root.