Put a white background behind images; add an "invert images in dark mode" display option
fixes #1148

The default theme puts a white background behind <img>, <svg> and <object type="image*"> tags.

There's a new display option "invert images in dark mode", which will invert the colours of all images in content areas when the background color is dark.

This also tidies up the `data-prefers-color-scheme` attribute so it's set to "light" or "dark" as appropriate when the "automatic" color scheme is chosen, instead of not being set at all.