trying to fix MathJax spacing
fixes #1146

I'm still not sure that this is the right way of copying the MathJax styles into the shadow DOM, but this at least fixes the obvious spacing problems we've seen.

The `MathJax.svgStylesheet` is now cloned into the shadow DOM, instead of just moved, and the `font-size-adjust` CSS rule doesn't apply to anything under `mjx-container`.