working on the Parser
see issue #677

Trying to get it ready to more easily be subclassed for different notations.

jme.Parser is now a Class.

Fixing a bug with closeMissingBrackets and addMissingArguments - the output list needs to know the length of the stack at the time the output was added, so that when we add an op output, we don't grab outputs from before an opening bracket preceding the op.

There are now sets of functions, which scopes know about.
The aim is to use these to restrict the functions available to students in their answers, and to support mutually exclusive notations in separate contexts.

TO DO:
o Make a BaseParser class, and a StandardParser that extends that with all the built-ins
    - decided against - too faffy
x Move the global JME settings like arity, associative, commutative into the parser.
    - jme-rules seems to make a lot of new TOp tokens without adding this information.
- Look at lazyOps - is there another mechanism that could be used for some functions?
- Sort out converseOps
- Think about how extensions add function sets.