escape the TeX rendering of a single underscore as a variable name
A single underscore is a valid variable name.

If the student enters an answer like `a_(b)` and "single-letter variable names" is turned on, it's interpreted as `a * _ * b`.
The underscore wasn't escaped when rendered in LaTeX, so the whole expression would be `a_b`, which looks like it's been interpreted correctly.

This changes `getNameInfo` so a variable name starting with an underscore always counts as `isLong` and is rendered in LaTeX with `\texttt` and the underscores are escaped.