jme.isDeterministic looks at the expression of lambda tokens
This isn't *strictly* correct: an anonymous function `a -> random(0,1)` *is* deterministic as an expression that produces a lambda token, but the output of the lambda is not deterministic.
To be really accurate, a lambda should only count as non-deterministic when it's applied, e.g. `i -> random(0,i)` is deterministic but `(i -> random(0,i))(4)` is not. But I'd have to remember that for every function that takes lambdas.

This was the easiest way of ensuring that jme.isDeterministic doesn't think that expressions using non-deterministic lambdas are deterministic, e.g. `map(i->random(i..2i), 0..3)`