add 'variable_generation_run_number' variable
When using `seedrandom` along with a variable testing condition, you have the problem that if the set of variables picked doesn't satisfy the condition, you're stuck.
This adds a variable `variable_generation_run_number`, which counts the number of times variable generation has been tried. You can use this as part of the random seed, to deterministically find a set of variables that works.