\var and \simplify only wrap the TeX in braces if it starts with a - or + character.
I hadn't documented what motivated the braces.
If you have `e^\var{x}`, then when `x` is negative, the number needs to be wrapped in braces so the minus sign and the digits are both rendered in superscript.

But when substituting something like a `<` relation into TeX, curly braces stop it getting the right spacing.

So the fudge is to only wrap in curly braces if the generated TeX starts with a - or + character.
The other option would be to make the author remember to put curly braces, which is easy to miss.